+++ CHANGELOG.md
@@ -6,21 +6,32 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-Implemented typed loop control with `twist!`. Become fully compatible with `?`.
+Implemented typed loop control with `twist!`. Make `terror!` fully compatible with `?`.
+
+### Breaking
+- Renamed `Return` method `valret` to `into_valret`
 
 ### Added
 - Implementation of `twist!` with `Looping` type and `anybox!`
 - Dirty utility macros `last!`, `next!`, `resume!`
 - `tear::extra` For importing everything in one fell swoop
-- `Try` trait for `ValRet` and `Moral`, under the "experimental" feature
+- Integrate with the`Try` trait under the "experimental" feature
 - enum accessors for `ValRet` and `Moral`
 - Maintenance badge on crates.io
+- `gut` for use with `terror!` in functions returning `Option<T>`
+- Ensure basic no-std
+- Add `combinators` feature flag to convert to `Either` anything that can convert to `Moral`
 
 ### Changed
 - `tear_if!` lets you use anything in its body, instead of just expressions.
 - Make `terror!` the preferred macro over `rip!`
 - `terror!` and `fear!` can now take both argument forms
 - `terror!` now uses the `From` trait, just like `?`
+- Rewrite README
+
+### Removed
+- `or_else` and `map_ret` methods on `ValRet`. Use `result().or_else()` or `result().map_err()`
+  instead
 
 ## [0.1.1] – 2020-05-19